✨ feat(gui): window override logic and tweak main UI layout window override helpers a dedicated module and adjust layout spacing to improve rendering and window proc setup.
- Refactor: relocate apply_window_override and installndproc into gui::windows::windowsoverrides and import them where needed.
-actor export get_hwnd_from_cc and set_ui_ctx from the newsoverrides module; remove local get_hwnd_from_cc helper from.rs to consolidate HWND handling.
- Init: call set_egui_ctx creating the eframe window so the override module has access to theui context.
 UI: remove FileAction enum and commented-out apply_style call to reduce dead code in mainwindow.rs.
- Layout: adjust column height and spacing (16px from the available height use::NONE + add_space, tweak item_spacing)
 to correct tab bar positioning and gaps.
-: add an extra bottom spacing after drawing thebar to avoid cramped content and re imports.

These organize platform-specific window override code into asingle module and refine UI spacing for more consistent layout and procedure installation.
