+++ src/gui/windows/mod.rs
@@ -5,3 +5,4 @@ pub mod mainwindow;
 pub mod mainwindow_imp;
 pub mod settings;
 pub mod navigation;
+pub mod windowsoverrides;
