+++ src/main.rs
@@ -2,20 +2,9 @@ mod core;
 mod gui;
 
 use crate::core::indexer::{WindowSizeMode, load_app_settings};
+use crate::gui::windows::windowsoverrides::{get_hwnd_from_cc, set_egui_ctx};
 use eframe::{NativeOptions, egui};
-use raw_window_handle::{HasWindowHandle, RawWindowHandle};
-use windows::Win32::Foundation::HWND;
 use windows::Win32::System::Com::{CoInitializeEx, COINIT_APARTMENTTHREADED};
-
-fn get_hwnd_from_cc(cc: &eframe::CreationContext<'_>) -> Option<HWND> {
-    let handle = cc.window_handle().ok()?;
-    let raw = handle.as_raw();
-
-    match raw {
-        RawWindowHandle::Win32(h) => Some(HWND(h.hwnd.get() as *mut std::ffi::c_void)),
-        _ => None,
-    }
-}
 
 fn main() -> eframe::Result<()> {
     unsafe {
@@ -55,6 +44,7 @@ fn main() -> eframe::Result<()> {
             cc.egui_ctx.set_fonts(fonts);
 
             let hwnd = get_hwnd_from_cc(cc);
+            set_egui_ctx(&cc.egui_ctx);
             Ok(Box::new(gui::MainWindow::new(hwnd)))
         }),
     )
